requirements.txt
flask
python-dotenv==1.0.1